hwtest-intg/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hwtest-intg"
version = "0.1.0"
description = "Integration tests and example code for hwtest hardware test automation"
requires-python = ">=3.10"
license = {text = "Unlicense"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: The Unlicense (Unlicense)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "hwtest-core",
    "hwtest-rack",
    "python-can>=4.0",
    "httpx>=0.24",
    "pyyaml>=6.0",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-timeout>=2.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0",
    "black>=24.0",
    "pylint>=3.0",
    "types-PyYAML",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
hwtest_intg = ["configs/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (require hardware)",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["can", "httpx"]
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.pylint.main]
py-version = "3.10"

[tool.pylint.format]
max-line-length = 100

[tool.pylint."messages control"]
disable = [
    "too-few-public-methods",
    "too-many-instance-attributes",
    "import-outside-toplevel",
]
